Making the library more linkable

+++ include/byte_array/consumers.hpp
@@ -98,19 +98,19 @@ namespace consumers {
        __m128i comp = _mm_load_si128(( __m128i*)compare);
        __m128i mptr = _mm_loadu_si128(( __m128i*)ptr); // TODO: Optimise to follow alignment
        __m128i mret = _mm_cmpeq_epi8(comp, mptr);
-       uint16_t found = _mm_movemask_epi8 (mret);
+       uint16_t found = uint16_t(_mm_movemask_epi8 (mret));
        
 
        while( (pos < str.size()) && ( !found ) ) {
          pos += 16;
          ptr += 16;
-         // TODO: Handle \"
+         // TODO: Handle \"x
          __m128i mptr = _mm_loadu_si128(( __m128i*)ptr);
          __m128i mret = _mm_cmpeq_epi8(comp, mptr);
-         found = _mm_movemask_epi8 (mret); 
+         found = uint16_t(_mm_movemask_epi8 (mret) ); 
        }
 
-       pos += __builtin_ctz(found);
+       pos += uint64_t(__builtin_ctz(found));
        
        if( pos >= str.size() )
          return -1;
