+++ include/protocols/swarm/node_details.hpp
@@ -1,6 +1,6 @@
 #ifndef NODE_DETAILS_HPP
 #define NODE_DETAILS_HPP
-
+#include"chain/transaction.hpp"
 #include "byte_array/referenced_byte_array.hpp"
 #include "protocols/swarm/entry_point.hpp"
 #include "mutex.hpp"
@@ -58,8 +58,8 @@ struct NodeDetails
   byte_array::ByteArray public_key;
   std::vector< EntryPoint > entry_points;
   
-  uint32_t default_port;
-  uint32_t default_http_port;
+  uint16_t default_port;
+  uint16_t default_http_port;
 }; 
 
 class SharedNodeDetails
@@ -91,13 +91,13 @@ public:
     mutex_.unlock();    
   }
 
-  uint32_t default_port()
+  uint16_t default_port()
   {
     std::lock_guard< fetch::mutex::Mutex > lock( mutex_ );
     return details_.default_port;    
   }
   
-  uint32_t default_http_port() 
+  uint16_t default_http_port() 
   {
     std::lock_guard< fetch::mutex::Mutex > lock( mutex_ );
     return details_.default_http_port;
